# swagger/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "description": "Creates a new user with the provided data. Returns the created user with all its details.",
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "description": "Partner api key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully created. Returns the created user data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. The provided data is invalid or missing required fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request. The user is not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedRequestDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. The user does not have permission to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenRequestDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-api-key-partner": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "description": "Returns a list of all users.",
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "description": "Partner api key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users successfully returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request when listing users.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request when listing users.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedRequestDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden access when listing users.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenRequestDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-api-key-partner": []
          }
        ],
        "summary": "List all users",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Fetches and returns the details of a user by its unique identifier (UUID).",
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "Partner api key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found and returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request when fetching user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request when fetching user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedRequestDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden access when fetching user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenRequestDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-api-key-partner": []
          }
        ],
        "summary": "Get a user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "description": "Updates the details of a user identified by its UUID. Returns the updated user.",
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "Partner api key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request when updating user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request when updating user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedRequestDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden access when updating user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenRequestDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-api-key-partner": []
          }
        ],
        "summary": "Update a user by ID",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "description": "Performs a delete on a user by its UUID. The user is not removed from the database, only marked as deleted.",
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "Partner api key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User successfully deleted."
          },
          "400": {
            "description": "Bad request when removing user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request when removing user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedRequestDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden access when removing user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenRequestDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-api-key-partner": []
          }
        ],
        "summary": "Delete a user by ID",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Backend Aziface Partner Assemble Example API",
    "description": "Documentação da API Backend Aziface Partner Assemble Example",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-api-key-partner": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Partner ApiKey"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "bearer",
        "description": "Autorization Bearer"
      }
    },
    "schemas": {
...
      "BadRequestDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message describing the bad request",
            "example": "Invalid input data"
          },
          "statusCode": {
            "type": "number",
            "description": "HTTP status code for the error",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Error type or name",
            "example": "Bad Request"
          }
        },
        "required": [
          "message",
          "statusCode",
          "error"
        ]
      },
      "UnauthorizedRequestDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message describing the unauthorized request"
          },
          "statusCode": {
            "type": "number",
            "description": "HTTP status code for the error",
            "example": 401
          },
          "error": {
            "type": "string",
            "description": "Error type or name",
            "example": "Unauthorized"
          }
        },
        "required": [
          "message",
          "statusCode",
          "error"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User's name",
            "example": "João Silva"
          },
          "assembleUserId": {
            "type": "string",
            "description": "External identifier of the user in other systems",
            "example": "external123"
          },
          "username": {
            "type": "string",
            "description": "User name",
            "example": "johnDoe"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "johnDoe"
          }
        },
        "required": [
          "name",
          "username",
          "password"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the user",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Name of the user",
            "example": "John Doe"
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "JohnDoe"
          },
          "assembleUserId": {
            "type": "string",
            "description": "External identifier of the user in other systems",
            "example": "external-123"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the user was created",
            "example": "2023-10-01T12:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the user was last updated",
            "example": "2023-10-01T12:00:00Z"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "description": "If null the user is not deleted, otherwise it is the timestamp when the user was deleted",
            "example": null
          }
        },
        "required": [
          "id",
          "username",
          "createdAt",
          "updatedAt"
        ]
      },
      "ForbiddenRequestDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message describing the forbidden request",
            "example": "You do not have permission to access this resource"
          },
          "statusCode": {
            "type": "number",
            "description": "HTTP status code for the error",
            "example": 403
          },
          "error": {
            "type": "string",
            "description": "Error type or name",
            "example": "Forbidden"
          }
        },
        "required": [
          "message",
          "statusCode",
          "error"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User's name",
            "example": "João Silva"
          },
          "assembleUserId": {
            "type": "string",
            "description": "External identifier of the user in other systems",
            "example": "external123"
          },
          "username": {
            "type": "string",
            "description": "User name",
            "example": "johnDoe"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "johnDoe"
          }
        }
      },
...
    }
  }
}